Don't eval all of protocol hook just to set up the renderer
